풀대출 이자 소수점까지입력,표시가능

+++ application/client/loan-client/src/pages/LoanPool.jsx
@@ -40,7 +40,7 @@ const LoanPool = () => {
         id,
         name: poolName,
         minDeposit: parseInt(minDeposit),
-        interestRate: parseInt(interestRate),
+        interestRate: parseFloat(interestRate),
         durationMonths: parseInt(duration),
         creatorAddress: profile.wallet_id,
         initialDeposit: parseInt(initialDeposit)
